move lb_state out of server.rs
