+++ sgl-pdlb/src/main.rs
@@ -1,7 +1,9 @@
 mod io_struct;
+mod lb_state;
 mod server;
 mod strategy_lb;
 
+use lb_state::{LBConfig, LBState};
 use server::{periodic_logging, startup};
 use tokio::signal;
 
@@ -15,7 +17,7 @@ fn main() -> anyhow::Result<()> {
         .map(|i| format!("233.233.233.233:{}", i))
         .collect::<Vec<String>>();
 
-    let lb_config = server::LBConfig {
+    let lb_config = LBConfig {
         host: "localhost".to_string(),
         port: 8080,
         policy: "random".to_string(),
@@ -24,7 +26,7 @@ fn main() -> anyhow::Result<()> {
         log_interval: 5,
         timeout: 600,
     };
-    let lb_state = server::LBState::new(lb_config.clone()).map_err(|e| anyhow::anyhow!(e))?;
+    let lb_state = LBState::new(lb_config.clone()).map_err(|e| anyhow::anyhow!(e))?;
     let ret: anyhow::Result<()> = actix_web::rt::System::new().block_on(async move {
         tokio::select! {
             _ = periodic_logging(lb_state.clone()) => {
